kfuzztest: add array annotation
There was not any mechanism in place in the KFuzzTest for distinguishing
between value pointers and array pointers in fuzz test inputs. This
information is difficult to parse without additional semantic
information on the context.

We introduce a new KFTF_ANNOTATE_ARRAY macro that annotates a field in a
KFuzzTest input struct as an array, removing this ambiguity.